+++ src/components/welcome-placeholder/WelcomePlaceholder.jsx
@@ -66,7 +66,7 @@ function WelcomePlaceholder(props) {
 							flexDirection: "column",
 							alignItems: positionText[props.position],
 							justifyContent: "space-between",
-							height: "70%",
+							height: "80%",
 							padding: "70px",
 						}}
 					>
