+++ src/pages/index.js
@@ -1,7 +1,7 @@
 import CreateTaskForm from "../components/create-task-form/CreateTaskForm";
 
 import RecoverPasswordWrapper from "./components/recoverPasswordWrapper/RecoverPasswordWrapper";
-import SingInPage from "./login-page/SignInExample";
+import SingInPage from "./login-page/SignIn";
 import RecoverPassword from "./recover-password/RecoverPassword";
 import RegisterPage from "./register-screen/RegisterPage";
 import ResetPassword from "./reset-password/ResetPassword";
